requirements.txt
langchain_openai
langchain_core
python-dotenv
streamlit
langchain_community
langchain
langserve
fastapi
uvicorn
sse_starlette
bs4
chromadb
faiss-cpu